fix: restructuring server
