--- data/users.json
@@ -1,8 +1,0 @@
-{
-  "users": {
-    "admin": "123",
-    "evandro": "123",
-    "gaita": "123"
-  }
-}
-
