+++ server/constants/users.py
@@ -1,0 +1,6 @@
+USERS = {
+  "adm": "123",
+  "evandro": "123",
+  "rafael": "123",
+  "vitor": "123"
+}
