+++ server/core/gatekeeper.py
@@ -1,0 +1,4 @@
+from constants.users import USERS
+
+def authenticate(username, password):
+    return username in USERS and USERS[username] == password
