--- server/video_stream.py
@@ -1,15 +1,0 @@
-import cv2
-import socket
-
-def start_video_stream(client_socket):
-    cap = cv2.VideoCapture(0)
-
-    while cap.isOpened():
-        ret, frame = cap.read()
-        if not ret:
-            break
-            
-        _, buffer = cv2.imencode('.jpg', frame)
-        client_socket.sendall(buffer.tobytes())
-
-    cap.release()
